Changed the classifier architechture (modules.py)
